Move d3d12 functions
